chore: Cleanup restructure
